requirements.txt
django>=1.8.2
djangorestframework>=3.2.2
chatterbot>=0.2.2
